using one determine_chain_endpoint_and_network
