+++ bittensor/core/subtensor.py
@@ -1,4 +1,5 @@
 import copy
+import sys
 from functools import lru_cache
 from typing import TYPE_CHECKING, Any, Iterable, Optional, Union, cast
